backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
werkzeug==2.0.1
pyjwt==2.1.0
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.0.2
prophet==1.1.4
python-dotenv==0.19.0
flask-cors==3.0.10 